Change caching strategy from tonic clients to tonic channels.
A Channel represents a connection to a remote host and is worth
re-using, while tonic clients are designed to be cheap to clone.

This is a preliminary rework.  In particular, it's probably not a good
idea that the shard assignment task creates a connection to every leader
node in the cluster early.  What if the shard assignment map changes
before any of the chennels are ever used?

Notable changes:
 * pool::ChannelPool - cloneable pool of channels, minimize contention
   between access to channels from different hosts

 * pool::ChannelPoolFactory - trait to inject a factory to create
   channels.  Currnetly only useable for unit testing.

 * GrpcClientCache - removed