Whitelist the external symbols we're willing to export from other crates
Motivation:  there is a work-in-progress official grpc crate being
developed.  At some point we might switch from tonic to that, so let's
check to ensure that there aren't any accidental tonic dependencies in
the public API.

cargo-check-external-types requires the nightly toolchain, so add it to
the build container.  When/if the APIs it uses are stabilized, we'll
want to remove it