batch_get: interleave network I/O with processing
The previous approach to batch_get would drive a limitted number of
shard requests in parallel, but buffer all results until everything
was completed before the client sees anything:

    futures::stream::iter(batch_get_futures)
        .buffer_unordered(self.config.max_parallel_requests())
        .flatten_unordered(None),

All of the I/O is driven in the `.buffer_unordered()` stage,
and the `.flatten_unordered()` merges results into a single stream.

This work-in-progress approach does the initial GRPC request in the
`buffer_unordered` stage, and then reads from the stream in the
`flatten_unordered` stage

The code is also generalized to handle more than one GRPC response on
the stream from each request.  This won't happen with the current server
implementation and would be hard to change without breaking the official
Go client.

TODO:
   * fix missing error context
   * maybe simplify in light of only one response per request
   * re-think concurrency limits