batch_get: interleave network I/O with processing
Re-work `batch_get` to separate sending the initial GRPC request to get
the resposne stream and retrieving the stream items.  The motivation here
is to avoid driving all network I/O upfront and buffering results before
the client can process it.

Also, we know that the server will send one ReadResponse per ReadRequest,
so we can avoid the complexity of reading multiple responses.  While Oxia
can change that behavior without violating the GRPC contract, the semantic
change would be a breaking one for the existing clients, so we're probably
good for a bit.  If not, it'll be a fun debug session and an easy fix.

The previous approach to batch_get drives a limitted number of
shard requests in parallel, but buffers all results until everything
was completed before the client sees anything:

    futures::stream::iter(batch_get_futures)
        .buffer_unordered(self.config.max_parallel_requests())
        .flatten_unordered(None),

All of the I/O is driven in the `.buffer_unordered()` stage,
and the `.flatten_unordered()` merges results into a single stream.

Maybe the concurrency limits should be revisited with this change?