Notifications re-work
Re-implement the notifcations state machine in terms of tokio_stream
StreamMap.  This reduces some of the boilerplate and lets the code
focus on keeping the state machine running.

To be done: support options for reconnecting when the remote servers
close connnections or start erroring.

Also, the unit tests don't work.  They pass, but they passed during
versions of this code that was known broken.  I've done some spot
testing using the cli to verify notifications are retrieved and
remote server shutdown is detected and handled apropriately if you
ignore the not-yet-implememented reconnect hooks