Update oxia proto to v0.15.1
Also, this commit changes the way the external oxia dependency is
handled.  We'll now just pull the needed bits into ext/oxia and use
it for both building the protobuf binding and the standalone server, if
needed.  Hopefully no more fighting with the go module system, especially
as upstream re-organizes it's tree.